+++ pytest_inmanta/parameters.py
@@ -63,6 +63,18 @@ inm_mod_in_place = BooleanTestParameter(
     group=param_group,
 )
 
+inm_no_strict_deps_check = BooleanTestParameter(
+    argument="--no-strict-deps-check",
+    environment_variable="INMANTA_NO_STRICT_DEPS_CHECKS",
+    usage=(
+        "Tell pytest-inmanta to run without using the deps check after module installation."
+        " When using the dependency check, an error is raised if there are conflicting requirements"
+        " when disabling the check, the less strict legacy behavior is used instead."
+    ),
+    default=False,
+    group=param_group,
+)
+
 
 # This is the legacy module repo option
 # TODO remove this in next major version bump