+++ pytest_inmanta/plugin.py
@@ -73,6 +73,7 @@ from .parameters import (
     inm_mod_in_place,
     inm_mod_repo,
     inm_no_load_plugins,
+    inm_no_strict_deps_check,
     inm_venv,
 )
 from .test_parameter import ParameterNotSetException, TestParameterRegistry
@@ -281,7 +282,9 @@ def project_factory(request: pytest.FixtureRequest) -> typing.Callable[[], "Proj
 
     def create_project(**kwargs: object):
         load_plugins = not inm_no_load_plugins.resolve(request.config)
+        no_strict_deps_check = inm_no_strict_deps_check.resolve(request.config)
         extended_kwargs: typing.Dict[str, object] = {
+            "no_strict_deps_check": no_strict_deps_check,
             "load_plugins": load_plugins,
             "env_path": env_dir,
             **kwargs,
@@ -605,6 +608,7 @@ class Project:
         project_dir: str,
         env_path: str,
         load_plugins: typing.Optional[bool] = True,
+        no_strict_deps_check: typing.Optional[bool] = False,
     ) -> None:
         """
         :param project_dir: Directory containing the Inmanta project.
@@ -613,6 +617,7 @@ class Project:
         """
         self._test_project_dir = project_dir
         self._env_path = env_path
+        self.no_strict_deps_check = no_strict_deps_check
         self._stdout: typing.Optional[str] = None
         self._stderr: typing.Optional[str] = None
         self.types: typing.Optional[typing.Dict[str, inmanta.ast.Type]] = None
@@ -663,7 +668,7 @@ class Project:
 
     def _create_project_and_load(self, model: str) -> module.Project:
         """
-        This method doesn the following:
+        This method does the following:
         * Add the given model file to the Inmanta project
         * Install the module dependencies
         * Load the project
@@ -678,12 +683,20 @@ class Project:
             module.Project.__init__
         )
         # The venv_path parameter only exists on ISO5+
+
         extra_kwargs_init = (
             {"venv_path": self._env_path}
             if "venv_path" in signature_init.parameters.keys()
             else {}
         )
-        test_project = module.Project(self._test_project_dir, **extra_kwargs_init)
+
+        if "strict_deps_check" in signature_init.parameters.keys():
+            extra_kwargs_init["strict_deps_check"] = not self.no_strict_deps_check
+
+        test_project = module.Project(
+            self._test_project_dir,
+            **extra_kwargs_init,
+        )
 
         ProjectLoader.load(test_project)