+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v 2.4.0 (?)
 Changes in this release:
 - Add fixture to change the Inmanta state dir to a writable location for the current user.
+- Add a new '--no-strict-deps-check' option to run pytest-inmanta using the legacy check on requirements.
+By default the new strict check of core will be used.
 
 # v 2.3.3 (2022-05-18)
 Changes in this release: